+++ GD_Assignment2/headers/Util.hpp
@@ -1,10 +1,6 @@
 #include <random>
 
-//inline float randInRange(float mn, float mx) {
-//	return (((float)rand() / RAND_MAX) * (mx - mn) + mn);
-//}
-
 template <typename T>
 inline T randInRange(T mn, T mx) {
-	return (T)(((double)rand() / RAND_MAX)* (mx - mn) + mn);
+	return (T)(((float)rand() / RAND_MAX)* (mx - mn) + mn);
 }
