+++ GD_Assignment2/src/Vec2.cpp
@@ -1,7 +1,7 @@
 #include "Vec2.hpp"
 #include <cmath>
 
-Vec2::Vec2(double x, double y)
+Vec2::Vec2(float x, float y)
 	:x(x), y(y)
 {
 }
@@ -26,12 +26,12 @@ Vec2 Vec2::operator-(const Vec2& other) const noexcept
 	return {x-other.x, y-other.y};
 }
 
-Vec2 Vec2::operator*(const double scale) const noexcept
+Vec2 Vec2::operator*(const float scale) const noexcept
 {
 	return {x*scale, y*scale};
 }
 
-Vec2 Vec2::operator/(const double scale) const noexcept
+Vec2 Vec2::operator/(const float scale) const noexcept
 {
 	return {x/scale, y/scale};
 }
@@ -48,13 +48,13 @@ void Vec2::operator-=(const Vec2& other) noexcept
 	y -= other.y;
 }
 
-void Vec2::operator*=(const double scale) noexcept
+void Vec2::operator*=(const float scale) noexcept
 {
 	x *= scale;
 	y *= scale;
 }
 
-void Vec2::operator/=(const double scale) noexcept
+void Vec2::operator/=(const float scale) noexcept
 {
 	x /= scale;
 	y /= scale;
@@ -62,17 +62,17 @@ void Vec2::operator/=(const double scale) noexcept
 
 void Vec2::normalize() noexcept
 {
-	double len = length();
+	float len = length();
 	*this = len > 0.01 ? *this/length() : 0;
 }
 
 Vec2 Vec2::normalize(const Vec2& vec) noexcept
 {
-	double len = vec.length();
+	float len = vec.length();
 	return len > 0.01 ? vec / len : 0;
 }
 
-double Vec2::length() const noexcept
+float Vec2::length() const noexcept
 {
 	return sqrt(x * x + y * y);
 }
